requirements.txt
streamlit
tensorflow==2.16.1
keras==3.0.5
numpy
pillow